Add helper method
