+++ tests/api/address.test.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
-import { createMocks, createRequest, createResponse } from 'node-mocks-http';
+import { createRequest, createResponse } from 'node-mocks-http';
 import handleAddress from '../../pages/api/address/index';
+import { interceptAuthenticationRequest } from './helpers/helper'
 
 jest.mock('axios');
 
@@ -10,17 +11,8 @@ const addressData = [{ addressLine1: 'Pitcairn house' }];
 
 describe('/api/address', () => {
   beforeEach(() => {
-    // Mock auth request
     axios.create = jest.fn(() => axios);
-
-    const mockJwt = 'token';
-
-    const response = {
-      status: 200,
-      data: mockJwt,
-    };
-
-    axios.post.mockImplementationOnce(() => Promise.resolve(response));
+    interceptAuthenticationRequest();
   });
 
   test('when called - returns address data', async () => {
