+++ tests/api/maintenance.test.js
@@ -1,22 +1,14 @@
 import axios from 'axios';
-import { createMocks, createRequest, createResponse } from 'node-mocks-http';
+import { createRequest, createResponse } from 'node-mocks-http';
 import handleMaintenance from '../../pages/api/maintenance/index';
+import { interceptAuthenticationRequest } from './helpers/helper'
 
 jest.mock('axios');
 
 describe('/api/maintenance', () => {
   beforeEach(() => {
-    // Mock auth request
     axios.create = jest.fn(() => axios);
-
-    const mockJwt = 'token';
-
-    const response = {
-      status: 200,
-      data: mockJwt,
-    };
-
-    axios.post.mockImplementationOnce(() => Promise.resolve(response));
+    interceptAuthenticationRequest();
   });
 
   test('when called - returns maintenance status', async () => {
